Fixed several problems with the abbreviation mode and ELSE.  The routine handle-abbrev-expansion used to reinsert the placeholder text that ELSE would remove when you type into the placeholder.  It would then also run the commands and avoid the expansion by an on purpose error.  This messed up VR mode since the error would stop processing of the changes.  Made a very similar routine called fix-else-abbrev-expansion, which would reinsert the placeholder text but not call the command.  It's now called just before the command is executed in vr-report-change.
Also made a temporary fix for the fact that expand-placeholder is not
a command that returns quickly, but requires user input.  Since it is
called before make-changes sends tick and change counts, vr.exe would
timeout.  The fix checks if the command it is about to execute is
expand-placeholder, and if it is report-changes simply doesn't do
anything.  Instead, make-changes, after sending tick and change
counts, checks if deferred-deferred-function is set, and in that case
executes the command.  Of course this doesn't work if
expand-placeholder is not the last command in the utterance, but on
the other hand that doesn't make sense, because expand-placeholder
requires input, so it should be the last.

It still isn't perfect, because ELSE uses before-change-functions to
delete the placeholder, which means that VR mode will get out of
sync.  I can't think of a simple way of taking care of this, when the
Emacs buffer changes before any characters have been entered, and
vr.exe thinks that all the characters have already been typed, it's
not straightforward to sort that out.