Smoothed out a number of rough edges concerning the continuous commands.  Most of the changes are in pbvElse, but there were a number of problems with the deferred-function variables not being cleared properly.  It still seems that sometimes expand-placeholder, which should be executed in make-changes, does not get run.  Instead, it will be run the next time you enter make-changes.  I have no idea why this happens.
Also fixed some problems with the execute-function function, stupid
Lisp stuff like whether the command sequence is a list or a vector,
and so on.

Also added the running of pre-command-hook and post-command-hook for
our "simulated" keyboard input.  I hope that this would have cured the
fact that macro definitions don't recognize the VR mode keystrokes,
but it didn't.  I don't know how macros read keystrokes when they're
being defined.