Inserted the necessary machinery for sending changes from the get-buffer-info call.  This only consists of setting vr-ignore-changes to queue up the requests, and then sending them at the end of vr-cmd-get-buffer-info.
Also Added a check in vr-report-change for vr-ignore-changes ==
unconditionally, which not totally unexpectedly always will ignore
that change and not send it to NaturallySpeaking.

The functions vr-zap-invisible and vr-restore-invisible will be
responsible for handling the text marked "vr-invisible".  It will be
deleted (as far as NaturallySpeaking knows) in get-buffer-info, and
then restored (to NaturallySpeaking) in end-recognition.  We'll see if
we can get it to work.