vr-zap-invisible and vr-restore-invisible now seem to work.  There were some issues with character positions and property searches running all the way to the end of the buffer, but they seem to be resolved.
Dictating with point in a word marked vr-invisible now sends the
proper commands to remove the invisible text from the
NaturallySpeaking buffer and then reinsert it at the end of the
utterance.  However, because the buffers disagree, character positions
will also disagree, and this needs to be taking care of.  I've defined
a variable vr-point-shift which contains the number of characters
point should be shifted.  I think this is enough, but we may have to
save both the length and the position of the chunks before and after
point.  Anyway, the communication protocol needs to be redesigned
again, because now we can't send the cursor position until after we've
made this change and we know what the position in the
NaturallySpeaking buffer actually should be.