Changed the get-buffer-info call so that selection and window position are sent AFTER buffer resynchronization and possible deletions due to invisible text.  That way, the positions will be "corrected" for the invisible text.
Vr-point-shift is a list containing 3 numbers, the beginning of the
deleted region, the position of point, and the end of the deleted
region.  This is necessary so that we can shift positions before and
after the invisible region correctly as well, something that is used
if the utterance is a selection command.

The position shifting is done in two functions vr-pfe and vr-pte (for
point-from-Emacs and point-to-Emacs, respectively).  It also takes
care of the "difference of 1" in the indexing between Lisp and C++.
This new functions are used in get-buffer-info, make-changes and
change-text, the three functions that can be called while the
invisible text is invisible.  It should probably be used in all cases,
since it's also a little more transparent than using 1+ and 1-...

Text insertion into the invisible regions now seems to work correctly,
with the restoration of the invisible text and the positioning of the
cursor doing the right thing.  There seems to be a problem with the
invisible text being overly "contagious", but that's likely me not
setting the non-stickiness of vr-invisible properly.