Split the function vr-find-invisible from vr-zap-invisible to make it more modular.  Find-invisible first runs the hook vr-invisible-hook, which makes it easy to hook in functions to check whether text should be made invisible (e.g. checking if we are in an else placeholder).
Change the way NaturallySpeaking is notified that invisible text has
been deleted.  What I used to do was delete it and then reinsert it,
but that will confuse other applications also checking for buffer
changes (again, else), so now I just generate the command manually and
send it to NaturallySpeaking.

Realized that there is a fundamental flaw in how I envisioned this
working.  If the invisible text, that NaturallySpeaking already thinks
doesn't exist, is deleted, we can NOT send this notification to
NaturallySpeaking again!  This is painful, because it means that we
have to look for invisible text in vr-report-change as well, and
prevent those changes from being sent!  I'm not quite sure how to code
this, but I'm afraid it won't be pretty... I think this speaks in
favor of not keeping two copies of the buffer at all, like Alain says
they do with VoiceCoder.