Implemented the check for invisible characters in vr-report-change. If vr-floating-invisible is non-nil, report-change calls vr-count-invisible, which searches the deleted string (from vr-modification-stack) for invisible characters.  Since the only interesting number in a delete command is the number of characters that was deleted, we then simply subtract the number of invisible characters in the number we send to NaturallySpeaking.
It seems to work now, but I realized that the only advantage of doing
this is that we get the correct spacing when we dictate the string
into a placeholder.  The buffers will still get out of sync, because
the placeholder is still deleted before the dictated text is
inserted.  This means that the command that was supposed to delete the
placeholder text instead actually will delete the text that was just
dictated, from NaturallySpeaking's point of view.

I guess this can be fixed by adjusting the positions, but then we need
a mechanism to detect when changes happen and figure out a way to
calculate the positions NaturallySpeaking would need to see.  It's
getting VERY messy at this point, this is just too complicated for it
to work with a current architecture... :-(